Prevent writing to binmask register when sync_nent goes high
This fixes a bug that occurred on the 2nd event if a tf_mem_bin
module is written to the same clock cycle sync_nent goes high. If
this happened, the binmask wouldn't be completely cleared for the
next BX, and TrackletProcessors would then read stale data.

There exists a guard against this for normal BX transitions, however
this guard wasn't used during the very first BX transition (when
sync_nent goes high).